WorkspaceState Updated Queued on Change
We have the en-queue 3000ms delay, however the way it's currently setup we could accidentally drop events.

Ex.
1 - Project Update
2 - Project Update (silenced as previous project is already being updated)
7 - Project Update

From timestamp (1+3)->7  our workspace failed to represent the correct workspace state. This could result in failed completions and other issues.

Additionally, This results in repeated calls (TS 1, 4, 7, etc.) to razor code-analysis (out of proc so a bit more expensive computation). This needs to be modified to cancel previous update if still in process, and proceed with new update.

Discovered while investigating: https://github.com/dotnet/aspnetcore/issues/21631

Fixes: https://github.com/dotnet/aspnetcore/issues/22277
