+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/ProjectWorkspaceStateGenerator.cs
@@ -1,12 +1,13 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System.Threading;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 
 namespace Microsoft.CodeAnalysis.Razor
 {
     internal abstract class ProjectWorkspaceStateGenerator : ProjectSnapshotChangeTrigger
     {
-        public abstract void Update(Project workspaceProject, ProjectSnapshot projectSnapshot);
+        public abstract void Update(Project workspaceProject, ProjectSnapshot projectSnapshot, CancellationTokenSource cts = default);
     }
 }
