+++ src/Razor/test/Microsoft.VisualStudio.LanguageServices.Razor.Test/ProjectSystem/TestProjectWorkspaceStateGenerator.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using Microsoft.CodeAnalysis.Razor;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.CodeAnalysis;
+using System.Threading;
 
 namespace Microsoft.VisualStudio.LanguageServices.Razor.Test
 {
@@ -23,7 +24,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Razor.Test
         {
         }
 
-        public override void Update(Project workspaceProject, ProjectSnapshot projectSnapshot)
+        public override void Update(Project workspaceProject, ProjectSnapshot projectSnapshot, CancellationTokenSource cts)
         {
             var update = (workspaceProject, projectSnapshot);
             _updates.Add(update);
